chore: 파일 구조 수정 & 주석 제거
chore: 파일 구조 수정, 일부 코드 주석 제거
