+++ src/pages/multiStudyRoom/multiStudyRoomContent/MultiStudyRoomContent.style.ts
