+++ src/pages/privateStudyRoom/PrivateStudyRoom.tsx
@@ -1,5 +1,5 @@
 import { SocketProvider } from '@/socket/SocketContext';
-import PrivateStudyRoomContent from './PrivateStudyRoomContent';
+import PrivateStudyRoomContent from './privateStudyRoomContent/PrivateStudyRoomContent';
 import { useParams } from 'react-router-dom';
 
 // Props로 공부방 정보를 받을 예정
